llm_evaluation_system/requirements.txt
google-generativeai
pip-system-certs
pip-tools
python-dotenv
pillow
requests
urllib3
numpy
pandas
PyMuPDF
plotly
sqlalchemy
psycopg2-binary
pgvector
langchain
langchain-google-genai
typing-extensions
python-dotenv
structlog
black
isort
flake8
mypy
Sphinx
sphinx-rtd-theme
pytest-cov